+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.6.2 - 2025-12-26
+## 0.7.0 - Unreleased
 
 ### Highlights